Final Media Change
Simplified some of the media query stuff. For some reason, it's actually acting really inconsistently. I refresh the page a few times, and even though none of the CSS changes, it seems to swap between a couple different ways of acting. Not sure what it is, but I'm not seeing it much anymore, so I think I'm done!